+++ Assets/CJY/Scripts/Quest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7aeaf34927767ab47894c5204d220517
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
