+++ Assets/CJY/Scripts/Quest/QuestManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cbe1439526523bd4da01bd3e8bcc26cf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
