+++ Assets/CJY/Scripts/Quest/QuestScrip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9a38a16a36b3c344ab46f6e6540b1ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
