+++ Assets/CJY/Scripts/Setting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cf7379b699d81204ebca63e08a27a69f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
