+++ Assets/CJY/Scripts/Setting/AudioManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41d7eaf1a4569d346ba5f535f64f292e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
