+++ Assets/CJY/Scripts/Setting/SceneAudioSettings.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[System.Serializable]
+public class SceneAudioSettings
+{
+    public string sceneName;
+    public float bgmVolume;
+    public float sfxVolume;
+}
+
