+++ Assets/CJY/Scripts/Setting/SceneAudioSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa7fa24cc2cd99c4ba1839e5d9ba2cf6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
