+++ Assets/CJY/Scripts/Setting/SettingControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ec4557cb213a834fb22a3b1e2ce6ea0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
