--- Assets/CJY/Scripts/Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e7dcc8e2790b90243a1aa256db495919
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
